Allow custom session id provider
