+++ packages/honeycomb-opentelemetry-web/package.json
@@ -103,6 +103,7 @@
     "@opentelemetry/exporter-trace-otlp-http": "~0.57.0",
     "@opentelemetry/instrumentation": "~0.57.0",
     "@opentelemetry/opentelemetry-browser-detector": "~0.57.0",
+    "@opentelemetry/web-common": "~0.57.0",
     "@opentelemetry/resources": "~1.30.0",
     "@opentelemetry/sdk-trace-base": "~1.30.0",
     "@opentelemetry/sdk-trace-web": "~1.30.0",
