--- packages/honeycomb-opentelemetry-web/src/session-id-span-processor.ts
@@ -1,62 +1,0 @@
-// This code will eventually be packaged upstream into a WebSDK package.
-// Once it is released as a package, this distro will depend directly on the upstream package.
-// https://github.com/open-telemetry/opentelemetry-js/pull/4325
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Span, SpanProcessor } from '@opentelemetry/sdk-trace-base';
-
-const SESSION_ID_BYTES = 16;
-const SHARED_CHAR_CODES_ARRAY = Array(32);
-
-export class SessionIdSpanProcessor implements SpanProcessor {
-  private _sessionId;
-  private _idGenerator = getIdGenerator(SESSION_ID_BYTES);
-
-  constructor() {
-    this._sessionId = this._idGenerator();
-  }
-
-  onStart(span: Span): void {
-    span.setAttribute('session.id', this._sessionId);
-  }
-
-  onEnd(): void {}
-
-  forceFlush(): Promise<void> {
-    return Promise.resolve();
-  }
-
-  shutdown(): Promise<void> {
-    return Promise.resolve();
-  }
-}
-
-function getIdGenerator(bytes: number): () => string {
-  return function generateId() {
-    for (let i = 0; i < bytes * 2; i++) {
-      SHARED_CHAR_CODES_ARRAY[i] = Math.floor(Math.random() * 16) + 48;
-      // valid hex characters in the range 48-57 and 97-102
-      if (SHARED_CHAR_CODES_ARRAY[i] >= 58) {
-        SHARED_CHAR_CODES_ARRAY[i] += 39;
-      }
-    }
-    return String.fromCharCode.apply(
-      null,
-      SHARED_CHAR_CODES_ARRAY.slice(0, bytes * 2),
-    );
-  };
-}
